MAINT: Moved well meshing helper functions
